+++ mobile/with-swift/example/Assets.xcassets/apple.imageset/Contents.json
@@ -17,8 +17,5 @@
   "info" : {
     "author" : "xcode",
     "version" : 1
-  },
-  "properties" : {
-    "template-rendering-intent" : "template"
   }
 }
